Add Project Record Error handling
Error handling included for Version number in adding a project record. Added an array of elements to chose from for the purpose in a project record as well as a dropdown select. In the fetch request i've added a new computed value that will check all the version numbers in the current project.