Fixed component of project
Deleted api file as it is now inside the composables. Moved the array of StageSelect to the app so it can be shared with the components that need it.  Progress in editing and deleting a project from the component. Changes to the display of buttons for edit and delete.  New composable for patching project created.